Added persistency. Changed diameter from int to float.

+++ TreeSpotter/src/com/cpsc310/treespotter/client/AdminTreeData.java
@@ -21,7 +21,7 @@ public class AdminTreeData extends ClientTreeData {
 		return null;
 	}
 
-	public int getDiameter() {
+	public float getDiameter() {
 		return 0;
 	}
